lab3  basic

+++ src/executors/delete_executor.cpp
@@ -18,6 +18,12 @@ std::shared_ptr<Record> DeleteExecutor::Next() {
   while (auto record = children_[0]->Next()) {
     // 通过 context_ 获取正确的锁，加锁失败时抛出异常
     // LAB 3 BEGIN
+    if (!context_.GetLockManager().LockTable(context_.GetXid(), LockType::IX, plan_->GetTableOid())) {
+      throw DbException("Cannot acquire lock");
+    }
+    if (!context_.GetLockManager().LockRow(context_.GetXid(), LockType::X, plan_->GetTableOid(), record->GetRid())) {
+      throw DbException("Cannot acquire lock");
+    }
     table_->DeleteRecord(record->GetRid(), context_.GetXid(), true);
     count++;
   }
